fix nil string snafus on ingest and disseminate Closes #462 Closes #180

+++ lib/daitss/proc/wip/from_aip.rb
@@ -162,7 +162,7 @@ module Daitss
       end
 
     ensure
-      FileUtils.rm_r tdir
+      FileUtils.rm_r tdir if tdir
     end
 
     # transfer sip descriptor
